// web/src/proto/dya/oled_status/oled_status.ts
export enum Orientation {
  ORIENTATION_LANDSCAPE = 0,
  ORIENTATION_PORTRAIT = 1,
  UNRECOGNIZED = -1,
}
// ...
export enum Animation {
  ANIMATION_OFF = 0,
  ANIMATION_CYBER_FACE = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface OledStatusConfig {
  orientation: Orientation;
  showCentralBattery: boolean;
  showPeripheralBatteries: boolean;
  showOutput: boolean;
  showLayer: boolean;
  showWpm: boolean;
  showModifiers: boolean;
  showHidIndicators: boolean;
  animation: Animation;
}

export interface Capabilities {
  centralBattery: boolean;
  peripheralBatteries: boolean;
  output: boolean;
  layer: boolean;
  wpm: boolean;
  modifiers: boolean;
  hidIndicators: boolean;
  portrait: boolean;
  landscape: boolean;
  animationCyberFace: boolean;
}

export interface GetConfigRequest {
}

export interface GetConfigResponse {
  config?: OledStatusConfig | undefined;
  capabilities?: Capabilities | undefined;
}

export interface SetConfigRequest {
  config?: OledStatusConfig | undefined;
}

export interface SetConfigResponse {
  config?: OledStatusConfig | undefined;
}

export interface ResetConfigRequest {
}

export interface ResetConfigResponse {
  config?: OledStatusConfig | undefined;
}

export interface ErrorResponse {
  message: string;
}

export interface Request {
  getConfig?: GetConfigRequest | undefined;
  setConfig?: SetConfigRequest | undefined;
  resetConfig?: ResetConfigRequest | undefined;
}

export interface Response {
  error?: ErrorResponse | undefined;
  getConfig?: GetConfigResponse | undefined;
  setConfig?: SetConfigResponse | undefined;
  resetConfig?: ResetConfigResponse | undefined;
}
// ...
export const GetConfigRequest: MessageFns<GetConfigRequest> = {
  encode(_: GetConfigRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetConfigRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetConfigRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): GetConfigRequest {
    return {};
  },

  toJSON(_: GetConfigRequest): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<GetConfigRequest>, I>>(base?: I): GetConfigRequest {
    return GetConfigRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetConfigRequest>, I>>(_: I): GetConfigRequest {
    const message = createBaseGetConfigRequest();
    return message;
  },
};
// ...
export const SetConfigRequest: MessageFns<SetConfigRequest> = {
  encode(message: SetConfigRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.config !== undefined) {
      OledStatusConfig.encode(message.config, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SetConfigRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSetConfigRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.config = OledStatusConfig.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SetConfigRequest {
    return { config: isSet(object.config) ? OledStatusConfig.fromJSON(object.config) : undefined };
  },

  toJSON(message: SetConfigRequest): unknown {
    const obj: any = {};
    if (message.config !== undefined) {
      obj.config = OledStatusConfig.toJSON(message.config);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SetConfigRequest>, I>>(base?: I): SetConfigRequest {
    return SetConfigRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SetConfigRequest>, I>>(object: I): SetConfigRequest {
    const message = createBaseSetConfigRequest();
    message.config = (object.config !== undefined && object.config !== null)
      ? OledStatusConfig.fromPartial(object.config)
      : undefined;
    return message;
  },
};
// ...
export const Request: MessageFns<Request> = {
  encode(message: Request, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.getConfig !== undefined) {
      GetConfigRequest.encode(message.getConfig, writer.uint32(10).fork()).join();
    }
    if (message.setConfig !== undefined) {
      SetConfigRequest.encode(message.setConfig, writer.uint32(18).fork()).join();
    }
    if (message.resetConfig !== undefined) {
      ResetConfigRequest.encode(message.resetConfig, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Request {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.getConfig = GetConfigRequest.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.setConfig = SetConfigRequest.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.resetConfig = ResetConfigRequest.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Request {
    return {
      getConfig: isSet(object.getConfig)
        ? GetConfigRequest.fromJSON(object.getConfig)
        : isSet(object.get_config)
        ? GetConfigRequest.fromJSON(object.get_config)
        : undefined,
      setConfig: isSet(object.setConfig)
        ? SetConfigRequest.fromJSON(object.setConfig)
        : isSet(object.set_config)
        ? SetConfigRequest.fromJSON(object.set_config)
        : undefined,
      resetConfig: isSet(object.resetConfig)
        ? ResetConfigRequest.fromJSON(object.resetConfig)
        : isSet(object.reset_config)
        ? ResetConfigRequest.fromJSON(object.reset_config)
        : undefined,
    };
  },

  toJSON(message: Request): unknown {
    const obj: any = {};
    if (message.getConfig !== undefined) {
      obj.getConfig = GetConfigRequest.toJSON(message.getConfig);
    }
    if (message.setConfig !== undefined) {
      obj.setConfig = SetConfigRequest.toJSON(message.setConfig);
    }
    if (message.resetConfig !== undefined) {
      obj.resetConfig = ResetConfigRequest.toJSON(message.resetConfig);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Request>, I>>(base?: I): Request {
    return Request.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Request>, I>>(object: I): Request {
    const message = createBaseRequest();
    message.getConfig = (object.getConfig !== undefined && object.getConfig !== null)
      ? GetConfigRequest.fromPartial(object.getConfig)
      : undefined;
    message.setConfig = (object.setConfig !== undefined && object.setConfig !== null)
      ? SetConfigRequest.fromPartial(object.setConfig)
      : undefined;
    message.resetConfig = (object.resetConfig !== undefined && object.resetConfig !== null)
      ? ResetConfigRequest.fromPartial(object.resetConfig)
      : undefined;
    return message;
  },
};

function createBaseResponse(): Response {
  return { error: undefined, getConfig: undefined, setConfig: undefined, resetConfig: undefined };
}

export const Response: MessageFns<Response> = {
  encode(message: Response, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.error !== undefined) {
      ErrorResponse.encode(message.error, writer.uint32(10).fork()).join();
    }
    if (message.getConfig !== undefined) {
      GetConfigResponse.encode(message.getConfig, writer.uint32(18).fork()).join();
    }
    if (message.setConfig !== undefined) {
      SetConfigResponse.encode(message.setConfig, writer.uint32(26).fork()).join();
    }
    if (message.resetConfig !== undefined) {
      ResetConfigResponse.encode(message.resetConfig, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Response {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.error = ErrorResponse.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.getConfig = GetConfigResponse.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.setConfig = SetConfigResponse.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.resetConfig = ResetConfigResponse.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Response {
    return {
      error: isSet(object.error) ? ErrorResponse.fromJSON(object.error) : undefined,
      getConfig: isSet(object.getConfig)
        ? GetConfigResponse.fromJSON(object.getConfig)
        : isSet(object.get_config)
        ? GetConfigResponse.fromJSON(object.get_config)
        : undefined,
      setConfig: isSet(object.setConfig)
        ? SetConfigResponse.fromJSON(object.setConfig)
        : isSet(object.set_config)
        ? SetConfigResponse.fromJSON(object.set_config)
        : undefined,
      resetConfig: isSet(object.resetConfig)
        ? ResetConfigResponse.fromJSON(object.resetConfig)
        : isSet(object.reset_config)
        ? ResetConfigResponse.fromJSON(object.reset_config)
        : undefined,
    };
  },

  toJSON(message: Response): unknown {
    const obj: any = {};
    if (message.error !== undefined) {
      obj.error = ErrorResponse.toJSON(message.error);
    }
    if (message.getConfig !== undefined) {
      obj.getConfig = GetConfigResponse.toJSON(message.getConfig);
    }
    if (message.setConfig !== undefined) {
      obj.setConfig = SetConfigResponse.toJSON(message.setConfig);
    }
    if (message.resetConfig !== undefined) {
      obj.resetConfig = ResetConfigResponse.toJSON(message.resetConfig);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Response>, I>>(base?: I): Response {
    return Response.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Response>, I>>(object: I): Response {
    const message = createBaseResponse();
    message.error = (object.error !== undefined && object.error !== null)
      ? ErrorResponse.fromPartial(object.error)
      : undefined;
    message.getConfig = (object.getConfig !== undefined && object.getConfig !== null)
      ? GetConfigResponse.fromPartial(object.getConfig)
      : undefined;
    message.setConfig = (object.setConfig !== undefined && object.setConfig !== null)
      ? SetConfigResponse.fromPartial(object.setConfig)
      : undefined;
    message.resetConfig = (object.resetConfig !== undefined && object.resetConfig !== null)
      ? ResetConfigResponse.fromPartial(object.resetConfig)
      : undefined;
    return message;
  },
};